+++ CHANGELOG.md
@@ -22,6 +22,22 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/) and this p
 ### Known Issues
 -->
 
+## v1.0.6
+
+### Added
+
+- Added additional translation strings
+
+###
+
+- Renamed `dealForInitative` to `dealForInitiative`. The old function will still be available until v1.1
+
+### Fixed
+
+- Fixed abbreviation for PP in german translation
+- The autolink Wildcards setting should now only apply to NPCs again
+- Fixed a few spelling mistakes in code comments
+
 ## v1.0.5
 
 ### Fixed